+++ packages/api-client/src/index.ts
@@ -1,0 +1,7 @@
+/* eslint-disable @typescript-eslint/no-namespace */
+
+export namespace SECFund {
+  export type Paths = import("./sec-fund").paths;
+  export type Components = import("./sec-fund").components;
+  export type Operations = import("./sec-fund").operations;
+}