+++ packages/api-client/ts-codegen.sh
@@ -1,0 +1,2 @@
+# SEC Fund Daily Info
+pnpm openapi-typescript data/sec-fund-daily-info.yaml -o src/sec-fund.d.ts