+++ apps/cron/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 0.1.0 is not noted here
 
+## [0.4.0] - 2025-12-07
+
+- use yahoo for thai stocks because it is easier to do than set, so we can get rid of 2GB thicc playwright
+- feat: price updater for thai mutal fund via SEC API
+- feat: price updater for cryptocurrencies via Binance(.th) API
+
 ## [0.3.0] - 2025-12-07
 
 - reduce sql values logging size by formatting as one item per line